Deepen listing intake and projection modules
